style(timecode): clear the clang-tidy floor for the timecode files
The timecode work introduced files that trip the repo's WarningsAsErrors
clang-tidy gate. The failures were latent because the macOS CI job's short
PR-test step ran before the tidy step and short-circuited it; with the
delayed-completion test now fixed, the tidy gate is reachable and must be clean.

Across recorder_engine/{ingest,timing} plus muxer/streamworker:
- performance-move-const-arg: drop no-op std::move on trivially-copyable
  optionals/queued packets (kept + NOLINT where the moved type is move-only
  in GPU builds).
- bugprone-implicit-widening-of-multiplication-result: widen before the
  multiply (per-day frame math, blue-frame memset sizes).
- bugprone-unchecked-optional-access: guard the just-taken evidence accesses.
- bugprone-exception-escape: contain teardown exceptions in the Muxer and
  StreamWorker destructors.
- bugprone-branch-clone: merge adjacent identical branches; NOLINT the
  order-dependent SEI fallbacks that cannot merge.
- bugprone-switch-missing-default-case / narrowing / member-init /
  clang-analyzer DivideZero: explicit default, cast, initializer, and guard.

No behavior change; the full unit suite stays green.

+++ recorder_engine/ingest/h26xseitimecode.cpp
@@ -295,6 +295,9 @@ H26xTimingDetail::TimecodeParseResult extractFromSeiRbsp(
             parsed = H26xTimingDetail::parseH264PicTiming(rbsp.mid(pos, int(payloadSize)),
                                                           *context->h264());
             handled = true;
+            // Distinct SEI payload cases legitimately share the "unsupported" outcome; they
+            // cannot merge because the specific handled cases must precede these fallbacks.
+            // NOLINTNEXTLINE(bugprone-branch-clone)
         } else if (codec == NativeVideoCodec::H264 && payloadType == 1) {
             sawUnsupported = true;
         } else if (payloadType == 4 &&
@@ -410,7 +413,8 @@ extractResult(const QByteArray& annexB, NativeVideoCodec codec, const H26xTiming
             continue;
         }
         if (codec == NativeVideoCodec::Hevc) {
-            const int layerId = ((uchar(nal[0]) & 0x01u) << 5) | (uchar(nal[1]) >> 3);
+            const int layerId =
+                static_cast<int>(((uchar(nal[0]) & 0x01u) << 5) | (uchar(nal[1]) >> 3));
             if (layerId != 0) {
                 sawUnsupported = true;
                 continue;
@@ -478,6 +482,9 @@ extractResult(const QByteArray& annexB, NativeVideoCodec codec, const H26xTiming
                 sawMalformed = true;
             }
         }
+        // The Malformed and the Hevc-mismatch cases share the sawMalformed outcome but are
+        // distinct, order-dependent status checks that cannot be merged.
+        // NOLINTNEXTLINE(bugprone-branch-clone)
         if (parsed.status == TimecodeParseStatus::Malformed) {
             sawMalformed = true;
         } else if (parsed.status == TimecodeParseStatus::Unsupported) {
